requirements.txt
asgiref==3.7.2
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.1
bs4==0.0.2
caer==2.0.8
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==4.4.2
distlib==0.3.8
Django==5.0.2
dnspython==2.7.0
filelock==3.13.1
Flask==3.0.3
fonttools==4.54.1
h11==0.14.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
MouseInfo==0.1.3
moviepy==1.0.3
mypy==1.10.0
mypy-extensions==1.0.0
networkx==3.4
numpy==1.26.4
opencv-contrib-python==4.10.0.82
outcome==1.3.0.post0
packaging==24.1
pika==1.3.2
pillow==10.2.0
pipenv==2023.11.15
platformdirs==4.1.0
proglog==0.1.10
psycopg2==2.9.10
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
pymongo==4.10.1
PyMsgBox==1.0.9
pyparsing==3.1.4
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.1
pywhatkit==5.4
requests==2.31.0
setuptools==69.0.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.0
websocket-client==1.8.0
Werkzeug==3.0.2
wikipedia==1.4.0
wsproto==1.2.0
